requirements.txt
Flask
openai
google-cloud-texttospeech
pydub
